authorization

+++ src/user/user.controller.ts
@@ -1,16 +1,29 @@
-import { Body, Controller, Delete, Get, Logger, Param, ParseIntPipe, Patch, Post, Query } from '@nestjs/common';
-import { UserService }                                                                    from './user.service';
-import { CreateUserDto }                                                                  from './dto/create-user.dto';
-import { UpdateUserDto }                                                                  from './dto/update-user.dto';
-import { ApiResponse, ApiTags }                                                           from '@nestjs/swagger';
-import { UserDto }                                                                        from './dto/user.dto';
-import { User }                                                                           from './entities/user.entity';
-import { CatchEntityErrorsHandler }                                                       from '../commons/filters/entity-errors-handler.filter';
-import { PageOptionsDto }                                                                 from '../commons/pagination/page-options.dto';
-import { PageDto }                                                                        from '../commons/pagination/page.dto';
-import { ApiPaginatedResponse }                                                           from '../commons/pagination/api-paginated-response.decorator';
-import { Serialize }                                                                      from '../commons/serialize.interceptor';
-import { plainToInstance }                                                                from 'class-transformer';
+import {
+  Body,
+  Controller,
+  Delete,
+  ForbiddenException,
+  Get,
+  Logger,
+  Param,
+  ParseIntPipe,
+  Patch,
+  Query,
+}                                   from '@nestjs/common';
+import { UserService }              from './user.service';
+import { UpdateUserDto }            from './dto/update-user.dto';
+import { ApiResponse, ApiTags }     from '@nestjs/swagger';
+import { UserDto }                  from './dto/user.dto';
+import { User }                     from './entities/user.entity';
+import { CatchEntityErrorsHandler } from '../commons/filters/entity-errors-handler.filter';
+import { PageOptionsDto }           from '../commons/pagination/page-options.dto';
+import { PageDto }                  from '../commons/pagination/page.dto';
+import { ApiPaginatedResponse }     from '../commons/pagination/api-paginated-response.decorator';
+import { Serialize }                from '../commons/serialize.interceptor';
+import { plainToInstance }          from 'class-transformer';
+import { ApiJwtAuth }               from './auth/decorators/api-jwt-auth.decorator';
+import { CurrentUser }              from './decorators/current-user.decorator';
+import { AuthService }              from './auth/auth.service';
 
 
 @ApiTags('user')
@@ -19,14 +32,8 @@ import { plainToInstance }                                                      
 export class UserController {
   private readonly logger = new Logger(UserController.name);
 
-  constructor(private readonly userService: UserService) {}
-
-  @Post()
-  @ApiResponse({ type: UserDto })
-  @Serialize(UserDto)
-  async create(@Body() createUserDto: CreateUserDto): Promise<User> {
-    return await this.userService.create(createUserDto);
-  }
+  constructor(private userService: UserService,
+              private authService: AuthService) {}
 
   @Get()
   @ApiPaginatedResponse(UserDto)
@@ -45,14 +52,26 @@ export class UserController {
     return await this.userService.findOneOrFail(id);
   }
 
+  @ApiJwtAuth()
   @Patch(':id')
   @Serialize(UserDto)
-  async update(@Param('id', ParseIntPipe) id: number, @Body() updateUserDto: UpdateUserDto): Promise<User> {
-    return await this.userService.update(id, updateUserDto);
+  async update(@Param('id', ParseIntPipe) id: number,
+               @Body() updateUserDto: UpdateUserDto,
+               @CurrentUser() currentUser: User): Promise<User> {
+    this.userService.checkOwnership(id, currentUser);
+
+    if ('isAdmin' in updateUserDto && !currentUser.isAdmin)
+      throw new ForbiddenException(`User ${currentUser.username} should be admin to change user.id = ${ id } admin status`)
+
+    return await this.authService.update(id, updateUserDto);
   }
 
+  @ApiJwtAuth()
   @Delete(':id')
-  async remove(@Param('id', ParseIntPipe) id: number): Promise<void> {
+  async remove(@Param('id', ParseIntPipe) id: number,
+               @CurrentUser() currentUser: User): Promise<void> {
+    this.userService.checkOwnership(id, currentUser);
     await this.userService.remove(id);
   }
+
 }
